+++ src/pages/_app.js
@@ -3,24 +3,39 @@ import React from "react";
 import track from "lib/tracking/track";
 import dispatch from "lib/tracking/dispatch";
 import withApolloClient from "lib/apollo/withApolloClient";
-import withTheme from "lib/theme/withTheme";
 import withShop from "containers/shop/withShop";
 import withTags from "containers/tags/withTags";
 import Layout from "components/Layout";
 import withMobX from "lib/stores/withMobX";
 import rootMobXStores from "lib/stores";
+import { MuiThemeProvider } from "@material-ui/core/styles";
+import CssBaseline from "@material-ui/core/CssBaseline";
+import JssProvider from "react-jss/lib/JssProvider";
+import getPageContext from "../lib/theme/getPageContext";
 
 @withApolloClient
 @withShop
 @withTags
 @withMobX
-@withTheme
 @track({}, { dispatch })
 export default class App extends NextApp {
+  constructor(props) {
+    super(props);
+    this.pageContext = getPageContext();
+  }
+
+  pageContext = null;
+
   componentDidMount() {
     // Fetch and update auth token in auth store
     rootMobXStores.authStore.setTokenFromLocalStorage();
     rootMobXStores.keycloakAuthStore.setTokenFromLocalStorage();
+
+    // Remove the server-side injected CSS.
+    const jssStyles = document.querySelector("#jss-server-side");
+    if (jssStyles && jssStyles.parentNode) {
+      jssStyles.parentNode.removeChild(jssStyles);
+    }
   }
 
   render() {
@@ -29,15 +44,26 @@ export default class App extends NextApp {
 
     return (
       <Container>
-        {
-          route === "/checkout" ? (
-            <Component shop={shop} {...rest} />
-          ) : (
-            <Layout shop={shop}>
-              <Component shop={shop} {...rest} />
-            </Layout>
-          )
-        }
+        <JssProvider
+          registry={this.pageContext.sheetsRegistry}
+          generateClassName={this.pageContext.generateClassName}
+        >
+          <MuiThemeProvider
+            theme={this.pageContext.theme}
+            sheetsManager={this.pageContext.sheetsManager}
+          >
+            <CssBaseline />
+            {
+              route === "/checkout" ? (
+                <Component pageContext={this.pageContext} shop={shop} {...rest} />
+              ) : (
+                <Layout shop={shop}>
+                  <Component pageContext={this.pageContext} shop={shop} {...rest} />
+                </Layout>
+              )
+            }
+          </MuiThemeProvider>
+        </JssProvider>
       </Container>
     );
   }
